gen 1/2 stat calculation in Rust
