+++ pkm_rs/src/gen9_sv/pk9.rs
@@ -11,12 +11,12 @@ use crate::traits::{HasSpeciesAndForm, PkmBytes};
 use pkm_rs_derive::IsShiny4096;
 use pkm_rs_resources::abilities::AbilityIndexBounded;
 use pkm_rs_resources::ball::Ball;
-use pkm_rs_resources::helpers;
 use pkm_rs_resources::metadata_source::MetadataSource;
 use pkm_rs_resources::moves::{MoveIndex, MoveSlots};
 use pkm_rs_resources::natures::NatureIndex;
 use pkm_rs_resources::ribbons::{ModernRibbon, ModernRibbonSet};
 use pkm_rs_resources::species::{FormMetadata, SpeciesForm, SpeciesMetadata};
+use pkm_rs_resources::stats;
 use pkm_rs_types::strings::SizedUtf16String;
 use pkm_rs_types::{
     AbilityNumber, BinaryGender, ContestStats, FlagSet, HyperTraining, Ivs, Language,
@@ -141,8 +141,8 @@ impl Pk9 {
         let level = species_and_form
             .get_species_metadata()
             .calculate_level(buf.exp());
-        let mint_nature = buf.mint_nature()?;
-        let stats = helpers::calculate_stats_modern(
+        let mint_nature: NatureIndex = buf.mint_nature()?;
+        let stats = stats::calculate_all_modern(
             MetadataSource::ScarletViolet,
             species_and_form,
             &buf.ivs(),
@@ -348,7 +348,7 @@ impl Pk9 {
     }
 
     pub fn calculate_stats(&self) -> Stats16Le {
-        helpers::calculate_stats_modern(
+        stats::calculate_all_modern(
             MetadataSource::ScarletViolet,
             self.species_and_form.0,
             &self.ivs,
